requirements.txt
requests==2.22.0
markdownify==0.5.1
Brotli==1.0.9
pdf2docx==0.5.6
pypandoc==1.11
python_docx==0.8.11
PyYAML==6.0
PyYAML==6.0.1
tqdm==4.62.3
tqdm==4.65.0
tqdm==4.63.0
